requirements.txt
Appium-Python-Client==5.0.0
attrs==25.3.0
certifi==2025.1.31
h11==0.16.0
idna==3.10
numpy==2.2.5
opencv-python==4.11.0.86
outcome==1.3.0.post0
pillow==11.2.1
PySocks==1.7.1
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
